fix: GUI moved to the app directory
