--- app/src/main/kotlin/Main.kt
@@ -1,3 +1,0 @@
-fun main(){
-    println("Test gradle")
-}
