+++ app/src/main/kotlin/app/presenter/treesPresenters/BsTreePresenter.kt
@@ -1,35 +1,34 @@
-package bstrees.presenter.treesPresenters
+package app.presenter.treesPresenters
 
-import bstrees.model.dataBases.serialize.strategies.AvlStrategy
+import bstrees.model.dataBases.serialize.strategies.BsStrategy
 import bstrees.model.dataBases.serialize.strategies.intToString
 import bstrees.model.dataBases.serialize.strategies.stringToInt
 import bstrees.model.dataBases.serialize.strategies.stringToString
 import bstrees.model.dataBases.serialize.types.SerializableTree
 
-
-object AvlTreePresenter {
+object BsTreePresenter {
     fun addNode(tree: SerializableTree, key: String, value: String): SerializableTree{
         return when(tree.keyType to tree.valueType){
             "String" to "String" -> {
-                val strategy = AvlStrategy(::stringToString, ::stringToString, ::stringToString, ::stringToString, tree.keyType, tree.valueType)
+                val strategy = BsStrategy(::stringToString, ::stringToString, ::stringToString, ::stringToString, tree.keyType, tree.valueType)
                 val bTree = strategy.deserializeTree(tree)
                 bTree.insert(key, value)
                 strategy.serializeTree(bTree, tree.name)
             }
             "Int" to "Int" -> {
-                val strategy = AvlStrategy(::intToString, ::intToString, ::stringToInt, ::stringToInt, tree.keyType, tree.valueType)
+                val strategy = BsStrategy(::intToString, ::intToString, ::stringToInt, ::stringToInt, tree.keyType, tree.valueType)
                 val bTree = strategy.deserializeTree(tree)
                 bTree.insert(key.toInt(), value.toInt())
                 strategy.serializeTree(bTree, tree.name)
             }
             "String" to "Int" -> {
-                val strategy = AvlStrategy(::stringToString, ::intToString, ::stringToString, ::stringToInt, tree.keyType, tree.valueType)
+                val strategy = BsStrategy(::stringToString, ::intToString, ::stringToString, ::stringToInt, tree.keyType, tree.valueType)
                 val bTree = strategy.deserializeTree(tree)
                 bTree.insert(key, value.toInt())
                 strategy.serializeTree(bTree, tree.name)
             }
             "Int" to "String" -> {
-                val strategy = AvlStrategy(::intToString, ::stringToString, ::stringToInt, ::stringToString, tree.keyType, tree.valueType)
+                val strategy = BsStrategy(::intToString, ::stringToString, ::stringToInt, ::stringToString, tree.keyType, tree.valueType)
                 val bTree = strategy.deserializeTree(tree)
                 bTree.insert(key.toInt(), value)
                 strategy.serializeTree(bTree, tree.name)
@@ -41,25 +40,25 @@ object AvlTreePresenter {
     fun deleteNode(tree: SerializableTree, key: String): SerializableTree {
         return when(tree.keyType to tree.valueType){
             "String" to "String" -> {
-                val strategy = AvlStrategy(::stringToString, ::stringToString, ::stringToString, ::stringToString, tree.keyType, tree.valueType)
+                val strategy = BsStrategy(::stringToString, ::stringToString, ::stringToString, ::stringToString, tree.keyType, tree.valueType)
                 val bTree = strategy.deserializeTree(tree)
                 bTree.delete(key)
                 strategy.serializeTree(bTree, tree.name)
             }
             "Int" to "Int" -> {
-                val strategy = AvlStrategy(::intToString, ::intToString, ::stringToInt, ::stringToInt, tree.keyType, tree.valueType)
+                val strategy = BsStrategy(::intToString, ::intToString, ::stringToInt, ::stringToInt, tree.keyType, tree.valueType)
                 val bTree = strategy.deserializeTree(tree)
                 bTree.delete(key.toInt())
                 strategy.serializeTree(bTree, tree.name)
             }
             "String" to "Int" -> {
-                val strategy = AvlStrategy(::stringToString, ::intToString, ::stringToString, ::stringToInt, tree.keyType, tree.valueType)
+                val strategy = BsStrategy(::stringToString, ::intToString, ::stringToString, ::stringToInt, tree.keyType, tree.valueType)
                 val bTree = strategy.deserializeTree(tree)
                 bTree.delete(key)
                 strategy.serializeTree(bTree, tree.name)
             }
             "Int" to "String" -> {
-                val strategy = AvlStrategy(::intToString, ::stringToString, ::stringToInt, ::stringToString, tree.keyType, tree.valueType)
+                val strategy = BsStrategy(::intToString, ::stringToString, ::stringToInt, ::stringToString, tree.keyType, tree.valueType)
                 val bTree = strategy.deserializeTree(tree)
                 bTree.delete(key.toInt())
                 strategy.serializeTree(bTree, tree.name)
