+++ app/src/main/kotlin/app/view/App.kt
@@ -1,28 +1,18 @@
-package bstrees.view
+package app.view
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.*
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.window.*
-import bstrees.model.dataBases.serialize.types.SerializableNode
-import bstrees.presenter.DataBasePresenter
-import bstrees.presenter.TreePresenter
-import bstrees.view.assets.ChildStack
-import bstrees.view.assets.ProvideComponentContext
-import bstrees.view.assets.Tree
-import bstrees.view.assets.TreeView
-import bstrees.view.screens.ChosingTypesScreen
-import bstrees.view.screens.HomeScreen
-import bstrees.view.screens.ScreenManager
-import bstrees.view.screens.TreeSreen
+import app.presenter.DataBasePresenter
+import app.presenter.TreePresenter
+import app.view.assets.ChildStack
+import app.view.assets.ProvideComponentContext
+import app.view.screens.ChosingTypesScreen
+import app.view.screens.HomeScreen
+import app.view.screens.ScreenManager
+import app.view.screens.TreeSreen
 import com.arkivanov.decompose.DefaultComponentContext
 import com.arkivanov.decompose.extensions.compose.jetbrains.stack.animation.fade
 import com.arkivanov.decompose.extensions.compose.jetbrains.stack.animation.plus
