+++ gradle.properties
@@ -6,3 +6,4 @@ neo4jDriverVersion=5.6.0
 sqliteJdbcVersion=3.34.0
 kotlinLoggingVersion=2.0.6
 slf4jVersion=1.7.29
+composeVersion=2.0.0-alpha-02
